automatically backed up by learn

+++ index.js
@@ -31,7 +31,7 @@ function dec(n) {
   }*/
 
 function makeInt (n) {
-  return parseFloat (n, 10);
+  return parseInt (n, 10);
 }
 
 /*function preserveDecimal(n) {
